modify the decision tree generator to allow deciding based upon an OutlierGroup directly, using await when necessary.
adapt the ImageCache to give PixelatedImages, be an actor, and have proper synchronization
during loading of an image, so multiple requests for the same filename don't result in
multiple concurrent loads of the same resource.

remove the existing decision trees, replace with a smaller version while the full
data is processed.